Code Condensing and Variable Renaming
Condensed the code that checks for number inputs to work in a for loop with a nested switch case that has only one case statement instead of using a switch case statement that uses 10 cases.

Renamed "tempResult" to "finalCalculation" and "tempCalcResult" to "tempCalculation" for better code readability.

The "C" button on the calculator's event listener is now a function that works within the on click event listener for the calculator buttons.

Removed console.log statements that were used older debugging.